A new version
Making the result symmetric and consistent with the Matlab.
It has higher precision and is faster, which may be a good version for academic use.
The instance used in it is the Fraunhofer diffraction. You can easily transform it into two-slit interference. And further, the Aharonove-Bohm effect, which is the goal of the research